python-face-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
face_recognition>=1.3.0
python-multipart>=0.0.6
Pillow>=10.0.0
